Revert "Saving work"
This reverts commit 168852a30ba2cc65522d517665f12c837b946127.

+++ test/unit_tests/unit_tests.cpp
@@ -412,7 +412,7 @@ auto write_string_to_file(Env &env, const std::string &filename, const std::stri
     if (offset < 0) {
         ASSERT_OK(env.file_size(filename, write_pos));
     } else {
-        write_pos = static_cast<std::size_t>(offset);
+        write_pos = offset;
     }
     ASSERT_OK(file->write(write_pos, buffer));
     ASSERT_OK(file->sync());
